replace std sleep with tokio sleep
this makes it more easy to shutdown a async task

+++ src/utils/config.rs
@@ -51,7 +51,7 @@ pub struct Mail {
     pub sender_pass: String,
     pub recipient: String,
     pub mail_level: String,
-    pub interval: i32,
+    pub interval: u64,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
